Move get_static_vram_buffer into gu_utils (#59)
* Move get_static_vram_buffer into sys

* Move into utils directory

* s/utils::gu/gu_utils/
