+++ examples/gu-background/src/main.rs
@@ -5,6 +5,7 @@
 
 use core::ffi::c_void;
 use psp::sys::{self, GuState, TexturePixelFormat, DisplayPixelFormat};
+use psp::gu_utils::get_static_vram_buffer;
 use psp::{BUF_WIDTH, SCREEN_WIDTH, SCREEN_HEIGHT};
 
 psp::module!("sample_gu_background", 1, 1);
@@ -54,33 +55,4 @@ fn psp_main() {
             sys::sceGuSwapBuffers();
         }
     }
-}
-
-fn get_memory_size(width: u32, height: u32, psm: TexturePixelFormat) -> u32 {
-    match psm {
-        TexturePixelFormat::PsmT4 => (width * height) >> 1,
-        TexturePixelFormat::PsmT8 => width * height,
-
-        TexturePixelFormat::Psm5650
-        | TexturePixelFormat::Psm5551
-        | TexturePixelFormat::Psm4444
-        | TexturePixelFormat::PsmT16 => {
-            2 * width * height
-        }
-
-        TexturePixelFormat::Psm8888 | TexturePixelFormat::PsmT32 => 4 * width * height,
-
-        _ => unimplemented!(),
-    }
-}
-
-unsafe fn get_static_vram_buffer(width: u32, height: u32, psm: TexturePixelFormat) -> *mut c_void {
-    static mut STATIC_OFFSET: u32 = 0;
-
-    let mem_size = get_memory_size(width, height, psm);
-    let result = STATIC_OFFSET as *mut _;
-
-    STATIC_OFFSET += mem_size as u32;
-
-    result
 }
