+++ psp/src/gu_utils.rs
@@ -1,0 +1,31 @@
+use core::ffi::c_void;
+use crate::sys::TexturePixelFormat;
+
+pub unsafe fn get_static_vram_buffer(width: u32, height: u32, psm: TexturePixelFormat) -> *mut c_void {
+    static mut STATIC_OFFSET: u32 = 0;
+
+    let mem_size = get_memory_size(width, height, psm);
+    let result = STATIC_OFFSET as *mut _;
+
+    STATIC_OFFSET += mem_size as u32;
+
+    result
+}
+
+fn get_memory_size(width: u32, height: u32, psm: TexturePixelFormat) -> u32 {
+    match psm {
+        TexturePixelFormat::PsmT4 => (width * height) >> 1,
+        TexturePixelFormat::PsmT8 => width * height,
+
+        TexturePixelFormat::Psm5650
+        | TexturePixelFormat::Psm5551
+        | TexturePixelFormat::Psm4444
+        | TexturePixelFormat::PsmT16 => {
+            2 * width * height
+        }
+
+        TexturePixelFormat::Psm8888 | TexturePixelFormat::PsmT32 => 4 * width * height,
+
+        _ => unimplemented!(),
+    }
+}
