+++ psp/src/lib.rs
@@ -31,6 +31,7 @@ pub mod debug;
 #[macro_use] mod vfpu;
 mod eabi;
 pub mod sys;
+pub mod gu_utils;
 
 #[cfg(not(feature = "stub-only"))] mod alloc_impl;
 #[cfg(not(feature = "stub-only"))] pub mod panic;
